requirements.txt
requests
ics
python-dotenv
tzdata
